vps/conversation-intelligence/requirements.txt
redis>=5.0.0
qdrant-client>=1.12.0
sentence-transformers>=3.3.0
openai>=1.58.0
networkx>=3.4.0
httpx>=0.28.0
pyyaml>=6.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
